+++ integrations/github-cloud/CHANGELOG.md
@@ -6,3 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- towncrier release notes start -->
+
+## 0.1.1-dev (2025-03-29)
+
+
+### Improvements
+
+- Added rate limiting to the GitHub Cloud integration.
+- Added support for Webhook Events and API calls
+- Added helpers functions
+
+## 0.1.0-dev (2025-03-22)
+
+
+### Features
+
+### Initial Integration:
+Created the GitHub Ocean integration.
+Added support for the following 
+- kinds:
+- Repository
+- Issue
+- Pull Request
+- Workflow